requirements.txt
# Core retrieval
faiss-cpu>=1.7.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.6.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.64.0
click>=8.0.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Streamlit UI
streamlit>=1.28.0
pillow>=9.0.0
